% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Paths.R
\name{all_shortest_paths_uwud_fast}
\alias{all_shortest_paths_uwud_fast}
\alias{distances_uwud_fast}
\title{All Pairs Shortest Paths (Unweighted & Undirected)}
\usage{
all_shortest_paths_uwud_fast(graph)

distances_uwud_fast(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
returns a symmetric "matrix" object where the i,jth entry is the
shortest path distance from v_i to v_j for all vertices in \code{graph}.
}
\description{
Computes all pairs shortest paths for an unweighted and undirected graph.
For graphs matching this description, this function is much faster than the
function \code{\link{distances}} in the igraph package.
}
\section{Functions}{
\itemize{
\item \code{distances_uwud_fast}: 
}}
\examples{
require(stats)
cb <- cluster_berenhaut(karate)
plot(cb)
}
\keyword{graphs}

