% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dissimilarity.R
\name{dissimilarity}
\alias{dissimilarity}
\title{Redwalk Dissimilarity Measure}
\usage{
dissimilarity(graph, nodes = V(graph), short_paths = NULL)
}
\arguments{
\item{graph}{an igraph object.}

\item{nodes}{a subset of the nodes in the graph to cluster, defaults to all
nodes (full community detection)}

\item{short_paths}{either a \eqn{|V|x|V|} symmetric dissimilarity matrix of
the vertices in \code{graph} or a \code{\link{dist}} object from the
\pkg{\link{stats}} package. If \code{NULL} (default), the shortest paths will
be calculated. Note: much faster if \code{diss} is provided and much faster
if \code{diss} is of class "matrix" rather than "dist".}
}
\value{
\code{dissimilarity} returns a dist object.
}
\description{
Compute the dissimilarity to be used in the redwalk clustering.
}
\examples{
require(stats, quietly = TRUE)
D <- dissimilarity(karate)
D

}
\keyword{cluster}
\keyword{dissimilarity}
\keyword{graph}
\keyword{redwalk}

