% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Distance.R
\name{mean_neighbor_distances_min}
\alias{get_d2_min}
\alias{mean_neighbor_distances_min}
\title{Mean Neighbor Distances (Minimum)}
\usage{
mean_neighbor_distances_min(graph, short_paths = NULL, as_dist = FALSE)
}
\arguments{
\item{graph}{an igraph object.}

\item{short_paths}{either a |V|x|V| symmetric matrix containing the shortest
path distances between the vertices in \code{graph} or a \code{\link{dist}}
object from the \code{\link{stats}} package. If \code{NULL} (default), the
shortest paths will be calculated. If shortest paths are available, this
functions runs at least 50\% faster. Counterintuitively, this function will
be faster if \code{short_paths} is of class "matrix" instead of class "dist".}

\item{as_dist}{if \code{FALSE} (default), will return a "matrix" object. If
\code{TRUE}, will return a "\code{\link{dist}}" object from \pkg{stats}.}
}
\value{
\code{mean_neighbor_distances_min} returns either a matrix or a dist
object representing dissimilarities between vertices in a graph.
}
\description{
For each vertex v_i in \code{graph}, this function calculates the average of
the shortest paths from each neighbor of v_i to each v_j in \code{graph}, but
the matrix that is returned will be the minimum value of the ij'th entry and
the ji'th entry, resulting in a symmetric matrix.
}
\examples{
mean_neighbor_distances_min(karate)
}
\keyword{cluster}
\keyword{graphs}

