\name{simple_network}
\docType{data}
\alias{simple_network}
\title{
  Simple Network
}
\description{
  A simple network for demonstration purposes. The network was created to show
  a reasonable 3 community partition.
}
\usage{
  data(simple_network)
}
\format{
  An undirected \code{\link{igraph}} graph object. The graph contains 15
  vertices and 28 edges with 3 communities.

  Graph attributes: \sQuote{name}.

  Vertex attributes: \sQuote{membership}, \sQuote{x}, \sQuote{y},
  \sQuote{color}, \sQuote{label.color}

  Edge attribute: \sQuote{color}.
}
\source{
  See reference below.
}
\references{
  Peter S. Barr, \emph{TODO}.
}
\keyword{datasets}
\keyword{networks}
