% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Paths.R
\name{all_shortest_paths_uwud_fast}
\alias{all_shortest_paths_uwud_fast}
\alias{distances_uwud_fast}
\title{All Pairs Shortest Paths (Unweighted & Undirected)}
\usage{
all_shortest_paths_uwud_fast(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
returns a symmetric "matrix" object where the \eqn{ij}th entry is the
shortest path distance from \eqn{v_i} to \eqn{v_j}, \eqn{d(v_i,v_j)}, for all
\eqn{v_i,v_j \in V}.
}
\description{
Computes all pairs shortest paths for an unweighted, undirected graph. For
graphs with these properties, this function is much faster than the function
\code{\link{distances}} in the \pkg{\link{igraph}} package.
}
\examples{
sp <- all_shortest_paths_uwud_fast(dolphins)
sp
all(sp == igraph::distances(dolphins))
}
\keyword{graphs}

